refactor: 모듈화 및 jsdoc추가
