+++ src/features/main/components/menu/index.ts
@@ -1,0 +1,1 @@
+export { default as Menu } from './Menu';
