+++ src/features/main/components/notice/NoticeBox.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import * as S from './NoticeBox.styles';
-import NoticeText from '@/features/main/components/notice/NoticeText';
+import { NoticeText } from '@/features/main/components/notice/index';
 import { NoticeBoxProps } from './Notice.types';
 
 function NoticeBox({ id, img, title, body, onClick }: NoticeBoxProps) {
