+++ src/features/main/components/notice/index.ts
@@ -1,0 +1,4 @@
+export { default as NoticeDetailCarousels } from './NoticeDetailCarousels';
+export { default as NoticeBody } from './NoticeBody';
+export { default as NoticeBox } from './NoticeBox';
+export { default as NoticeText } from './NoticeText';
