+++ src/pages/index.ts
@@ -8,8 +8,8 @@ import User from '@/pages/User';
 import TimeTable from '@/pages/performance/TimeTable';
 import PerformanceDetail from '@/pages/performance/PerformanceDetail';
 import LostPost from './main/LostPost';
-import Notice from '@/pages/main/Notice';
-import NoticeDetail from '@/pages/main/NoticeDetail';
+import Notice from '@/pages/main/notice/Notice';
+import NoticeDetail from '@/pages/main/notice/NoticeDetail';
 import Lost from './main/Lost';
 
 export {
