+++ src/pages/main/Main.tsx
@@ -1,7 +1,7 @@
 import { NavBar } from '@/components/nav-bar';
 import * as S from './Main.styles';
 import { EventCarousels } from '@/features/main/components/carousels';
-import Menu from '@/features/main/components/menu/Menu';
+import { Menu } from '@/features/main/components/menu/index';
 
 export default function Main() {
   return (
