+++ src/pages/main/notice/Notice.styles.ts
