+++ src/pages/main/notice/NoticeDetail.styles.ts
