+++ src/pages/main/notice/index.ts
@@ -1,0 +1,2 @@
+export { default as Notice } from './Notice';
+export { default as NoticeDetail } from './NoticeDetail';
